Adding error-handling
Quite a few changes here.

Added an errortype called LispError which is an instance of Error, so GHC's built-in error handling functions can be used.
A curried typealias `ThrowsError` to an `Either LispError`. So methods returning `ThrowsError Type` returns an `Either LispError Type`. (Which should be self-explanatory)
The rest of the changes are basically changing the signatures and methods from `LispVal` to `ThrowsError LispVal` using their Monadic form. (Because Either is in instance of a Monad)